+++ ci/bin/verify-examples.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 DIR=${1:-examples}
+_PWD=$PWD
 
 source $(dirname $0)/terraform.sh
 
@@ -16,6 +17,9 @@ for example in ${EXAMPLES} ; do
     echo no tf files
     exit 1
   fi
+  cd ${example}
   verifyModulesAndPlugins ${example}
+  formatCheck ${example}
   validate ${example}
+  cd $_PWD
 done