+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## 4.2.0 - 2019-08-18
+
+- Added
+  - Option to disable account id is used in bucket name #113 @Glen-Moonpig 
+  - Cancel sport instances during destroy for example default and public.
+- Changed:
+  - Fixed typos #118 @mpsq 
+
+
 ## 4.1.0 - 2019-08-05
 
 - Changed:
@@ -45,7 +55,7 @@ Module is available as Terraform 0.11 module, pin module to version 3.x. Please 
 
 ## 3.7.0 - 2019-07-17
 - Changed
-  - Creation of multiple instances of the runner is now supported. Cache is therefor moved to an internal module. Pleas see the example `runner-public` for a concrete sample. The change should have no effect if you apply the state migration script `migragations/migration-state-3.7.x.sh`.
+  - Creation of multiple instances of the runner is now supported. Cache is therefore moved to an internal module. Pleas see the example `runner-public` for a concrete sample. The change should have no effect if you apply the state migration script `migragations/migration-state-3.7.x.sh`.
   - Examples are more generic by removing the time zone and AZ zone to variables. @@theBenForce
 
 ## 3.6.0 - 2019-07-04
@@ -63,7 +73,7 @@ Module is available as Terraform 0.11 module, pin module to version 3.x. Please 
 
 ## 3.7.0 - 2019-07-17
 - Changed
-  - Creation of multiple instances of the runner is now supported. Cache is therefor moved to an internal module. Pleas see the example `runner-public` for a concrete sample. The change should have no effect if you apply the state migration script `migragations/migration-state-3.7.x.sh`.
+  - Creation of multiple instances of the runner is now supported. Cache is therefore moved to an internal module. Pleas see the example `runner-public` for a concrete sample. The change should have no effect if you apply the state migration script `migragations/migration-state-3.7.x.sh`.
   - Examples are more generic by removing the time zone and AZ zone to variables. @@theBenForce
 
 ## 3.6.0 - 2019-07-04
@@ -82,7 +92,7 @@ Module is available as Terraform 0.11 module, pin module to version 3.x. Please 
 ## [3.4.0] - 2019-06-06
 - Changed:
   - Update default runner type, GitLab runner version, and versions in examples.
-  - Buffix #75 runner is not reachable when runners_use_private_address = false 
+  - Buffix #75 runner is not reachable when runners_use_private_address = false
   - Buffix - Missing typ - @Orkin
   - Bugfix #72 - Detect and retry when docker machine download fails eliasdorneles
   - Bugfix #68 - add count to prevent resource creation failing  @philippefuentes
@@ -200,7 +210,8 @@ Module is available as Terraform 0.11 module, pin module to version 3.x. Please 
 - Update default AMI's to The latest Amazon Linux AMI 2017.09.1 - released on 2018-01-17.
 - Minor updates in the example
 
-[Unreleased]: https://github.com/npalm/terraform-aws-gitlab-runner/compare/4.1.0...HEAD
+[Unreleased]: https://github.com/npalm/terraform-aws-gitlab-runner/compare/4.2.0...HEAD
+[4.2.0]: https://github.com/npalm/terraform-aws-gitlab-runner/compare/4.1.0...4.2.0
 [4.1.0]: https://github.com/npalm/terraform-aws-gitlab-runner/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/npalm/terraform-aws-gitlab-runner/compare/3.10.0...4.0.0
 [3.10.0]: https://github.com/npalm/terraform-aws-gitlab-runner/compare/3.9.0...3.10.0